Improved Prerequisites - API CHANGE
Prerequisites.cs
1) Changed internal representation of how many prerequisites are
satisfied to an int. This way all prerequisites only have to be looped
through on initialization of a prerequisite. After which, callbacks will
handle updating the int.
2) Changed cbSatisfied to cbSatisfiedChanged<bool> so that subscribers
can know when a prerequisite is no longer satisfied in addition to being
satisfied intially.
3) Removed getter for the internal List<Prerequisite> to prevent other
classes (even derived ones) from adding a prerequisite without
registering callbacks and whatnot
4) Added API for registering and unregistering Prerequisites at runtime.
5) Every Prerequiste will  register a callback to its prerequisites to
stay updated.
6) Added protected void SelfSatisfied so that a Prerequisite could
itself be satisfied, but be held back by its prerequisites. Defaults to
true but changeable in the editor and at runtime.

Because of these changes, a Prerequisite changing from satisfied to
unsatisfied will cause a chain reaction of satisfied changes to all
Prerequistes that depend on it.